+++ test/assets/tiles/New2.png
