+++ test/assets/tiles/New3.png
